robotic_arm/pincherx100/pick_place_system/requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
PyYAML>=6.0
flask>=3.0.0
dynamixel-sdk>=3.8.0
pyserial>=3.5

